codemetrics/requirements.txt
asgiref==3.8.1
astroid==3.3.9
colorama==0.4.6
dill==0.4.0
Django==4.2.17
djangorestframework==3.16.0
flake8==7.2.0
isort==6.0.1
mando==0.7.1
mccabe==0.7.0
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
pylint==3.3.6
radon==6.0.1
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2025.2
